requirements.txt
# Core Flask framework
Flask==3.0.3
Flask-WTF==1.2.1
Werkzeug==3.0.3

# Environment and configuration
python-dotenv==1.0.1

# AI and machine learning
google-generativeai>=0.8.0

# Image processing
Pillow==10.2.0

# Utilities
requests==2.31.0
markupsafe==2.1.3